+++ requirements.txt
@@ -2,3 +2,4 @@ requests>=2.7.0
 beautifulsoup4>=4.3.2
 six>=1.10.0
 pymongo==3.3.0
+Flask==0.11.1
